fix: removed unused constant and moved the common function in webex_utils.py
