+++ package/bin/webex_constants.py
@@ -15,7 +15,6 @@ _GET_MEETING_QUALITIES =  "meeting/qualities"
 _GET_DETAILED_CALL_HISTORY = "cdr_feed"
 _SECURITY_AUDIT_EVENTS_ENDPOINT = "admin/securityAudit/events"
 _FEDRAMP_BASE_URL = "api-usgov.webex.com"
-_Webex_Contact_Center_BASE_URL_PREFIX = "api.wxcc"
 _Webex_Contact_Center_BASE_URL  = "api.wxcc-{region}.cisco.com"
 
 UNAUTHORIZED_STATUS = 401
